Add: logic for scripting challenge tabs
